fedextrackapi/requirements.txt
absl-py==2.1.0
asgiref==3.8.0
astunparse==1.6.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
contourpy==1.2.1
cramjam==2.8.2
crispy-bootstrap4==2024.1
cvzone==1.6.1
cycler==0.12.1
Django==4.2.11
django-admin-honeypot-updated-2021==1.2.0
django-crispy-forms==2.1
django-ipware==6.0.4
django-mongodb-engine==0.6.0
djangorestframework==3.15.1
djangotoolbox==1.8.0
djongo @ file:///C:/sample_proj/djongo-1.3.6
dnspython==2.6.1
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.63.0
h5py==3.11.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
jax==0.4.26
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mediapipe==0.10.11
ml-dtypes==0.3.2
mongoengine==0.28.2
mysqlclient==2.2.1
namex==0.0.8
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.3.0
protobuf==3.20.3
pycparser==2.22
Pygments==2.17.2
pymongo==4.6.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-ipware==2.0.2
python-snappy==0.7.1
pytz==2022.2.1
requests==2.31.0
rich==13.7.1
scipy==1.13.0
six==1.16.0
sounddevice==0.4.6
sqlparse==0.4.4
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
typing_extensions==4.10.0
tzdata==2021.5
urllib3==2.2.1
Werkzeug==3.0.2
winkerberos==0.10.0
wrapt==1.16.0
zipp==3.18.1
